backend/pyproject.toml
[tool.poetry]
name = "weightracker"
version = "0.1.0"
description = ""
authors = ["David <david@sulaco.co.uk>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
psycopg2-binary = "^2.9.3"
django = "^4.2.7"
gunicorn = "^21.2.0"
djangorestframework = "^3.14.0"
model-bakery = "^1.17.0"
google-auth = "^2.29.0"
google-api-python-client = "^2.128.0"
django-cors-headers = "^4.3.1"
requests = "^2.32.2"

[tool.poetry.dev-dependencies]
pydevd-pycharm = "^222.3345.40"
watchdog = {extras = ["watchmedo"], version = "^2.3.0"}

[tool.black]
include = '\.py?$'
line-length = 79

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 79
skip_gitignore = true
sections='''
FUTURE,
STDLIB,
THIRDPARTY,
FIRSTPARTY,
LOCALFOLDER,
'''

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
